Add "Mark/Close Stale Issues/PRs" GitHub Action
This GitHub action [1] automates marking and closing stale issues/PRs.
Issues/PRs which have seen no activity for more than 180 days are
marked with the 'stale' label and a comment is added informing the the
people subscribed to the issue that unless the 'stale' label is
removed, or a new comment is added to the issue/PR it will be
automatically closed in 30 days.

This automation will help keep the number of issues/PRs manageable.

[1] https://github.com/actions/stale

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>